refactor: replace sidebar action notifications with direct calls and @FocusedValue
